requirements.txt
bcrypt==4.2.1
blinker==1.9.0
click==8.1.8
dateparser==1.2.1
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
pillow==11.1.0
python-dateutil==2.9.0.post0
pytz==2025.1
regex==2024.11.6
six==1.17.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
tzlocal==5.3
Werkzeug==3.1.3
